requirements.txt
requests~=2.32.3
loguru~=0.7.2
python-dotenv~=1.0.1